sass: Update post views
The content of each post in now limited to three lines
in the short view and in a long view has a scroll element.

The image height is also set in the same manner.

The posts are now ordered as title, image and then content.
The margins and padding are adjusted accordingly.

Signed-off-by: Alan Gaurav <alangaurav@gmail.com>